Bug 971162 - [B2G][Clock][Alarm] Clock alarm has vibration when alarm vibration is set to off. r=gaye
