+++ apps/clock/test/unit/mocks/mock_shared/js/l10n.js
@@ -26,6 +26,7 @@
       }
       return key;
     },
+    ready: function(cb) { setTimeout(cb); },
     translate: function() {},
     localize: function() {},
     DateTimeFormat: DateTimeFormat
